GH-#2: Renaming of entities
