+++ src/main/java/com/gamedoora/model/dao/Comment.java
@@ -1,6 +1,5 @@
 package com.gamedoora.model.dao;
 
-import java.io.Serializable;
 import java.util.Date;
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -14,7 +13,7 @@ import javax.persistence.TemporalType;
 
 @Entity
 @Table(name = "comments")
-public class Comments implements Serializable {
+public class Comment {
 
     @Id
     @GeneratedValue
