+++ src/main/java/com/gamedoora/model/dao/Role.java
@@ -1,0 +1,48 @@
+package com.gamedoora.model.dao;
+
+import java.io.Serializable;
+import javax.persistence.Basic;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+/**
+ * @author rajsh
+ */
+
+@Entity
+public class Role implements Serializable {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+    @Basic
+    private String name;
+    @Basic
+    private String description;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+}
