+++ src/main/java/com/gamedoora/model/dao/Tag.java
@@ -12,28 +12,22 @@ import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 
 @Entity
-@Table(name = "tag_entities")
-public class TagEntities implements Serializable {
+@Table(name = "tags")
+public class Tag implements Serializable {
 
     @Id
     @GeneratedValue
-    @Column(name = "id", table = "tag_entities", nullable = false)
+    @Column(name = "id", table = "tags", nullable = false)
     private Integer id;
     @Basic(optional = false)
-    @Column(name = "tag_id", table = "tag_entities", nullable = false)
-    private int tagId;
+    @Column(name = "name", table = "tags", nullable = false, length = 128)
+    private String name;
     @Basic(optional = false)
-    @Column(name = "entity_id", table = "tag_entities", nullable = false)
-    private int entityId;
-    @Basic(optional = false)
-    @Column(name = "entity_type", table = "tag_entities", nullable = false)
-    private String entityType;
-    @Basic(optional = false)
-    @Column(name = "created_at", table = "tag_entities", nullable = false)
+    @Column(name = "created_at", table = "tags", nullable = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date createdAt;
     @Basic(optional = false)
-    @Column(name = "updated_at", table = "tag_entities", nullable = false)
+    @Column(name = "updated_at", table = "tags", nullable = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date updatedAt;
 
@@ -45,28 +39,12 @@ public class TagEntities implements Serializable {
         this.id = id;
     }
 
-    public int getTagId() {
-        return tagId;
+    public String getName() {
+        return name;
     }
 
-    public void setTagId(int tagId) {
-        this.tagId = tagId;
-    }
-
-    public int getEntityId() {
-        return entityId;
-    }
-
-    public void setEntityId(int entityId) {
-        this.entityId = entityId;
-    }
-
-    public String getEntityType() {
-        return entityType;
-    }
-
-    public void setEntityType(String entityType) {
-        this.entityType = entityType;
+    public void setName(String name) {
+        this.name = name;
     }
 
     public Date getCreatedAt() {
