+++ src/main/java/com/gamedoora/model/dto/GdLogin.java
@@ -1,0 +1,20 @@
+package com.gamedoora.model.dto;
+
+public class GdLogin {
+    private String token;
+
+    /**
+     * @return the token
+     */
+    public String getToken() {
+        return token;
+    }
+
+    /**
+     * @param token the token to set
+     */
+    public void setToken(String token) {
+        this.token = token;
+    }
+    
+}
